requirements.txt
amqp==2.4.2
Babel==2.6.0
billiard==3.6.0.0
cairocffi==1.0.2
CairoSVG==2.3.1
celery==4.3.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==1.11.8
django-crispy-forms==1.7.2
django-jet==1.0.8
django-paypal==1.0.0
django-rest-framework==0.1.0
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
flower==0.9.3
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
jet-django==0.5.9
kombu==4.5.0
Pillow==5.4.1
psycopg2==2.8.2
pycparser==2.19
PyJWT==1.7.1
Pyphen==0.9.5
pytz==2018.9
requests==2.21.0
six==1.12.0
tinycss2==1.0.2
tornado==5.1.1
transliterate==1.10.2
urllib3==1.24.3
vine==1.3.0
WeasyPrint==47
webencodings==0.5.1
whitenoise==4.1.2
